Structure fixed - have to fix php file for better site safety
