+++ php_functions/insertTrainingWithDate.php
@@ -1,18 +1,5 @@
 <?php
-function connectToDb() {
-  $servername = "127.0.0.1";
-  $username = "bartek";
-  $password = "gymsitedb321";
-  $dbname = "gymsitedatabase_final3";
-
-  // Tworzenie połączenia
-  $conn = mysqli_connect($servername, $username, $password, $dbname);
-  // Sprawdzanie połączenia
-  if (!$conn) {
-    exit("Połączenie nieudane: " . mysqli_connect_error());
-  }
-  return $conn;
-}
+require('db.php');
 
 if (isset($_POST['WstawDate'])) {
   $conn = connectToDb();
@@ -44,3 +31,4 @@ if (isset($_POST['WstawDate'])) {
   mysqli_close($conn);
   header("Location: http://localhost/Gym_Site/index.php"); 
 }
+?>
