+++ php_functions/login.php
@@ -30,4 +30,4 @@ if (!empty($_POST['login']) && !empty($_POST['password'])) {
         echo $error;
     }
 }
-
+?>
