+++ php_functions/logout.php
@@ -4,3 +4,4 @@ session_start();
 unset($_SESSION['user']);
 session_destroy();
 header("Location: http://localhost/Gym_Site/index.php");
+?>
